Issue 281: add application and infraestructure layout

+++ Poliedro.Eds.Domain/PosOfSale/Entities/PosOfSaleEntity.cs
@@ -9,7 +9,7 @@ namespace Poliedro.Eds.Domain.PointOfSale.Entities
     public class PosOfSaleEntity : AuditableEntity
     {
         [Key]
-        public ulong IdPos { get; set; }
+        public int IdPos { get; set; }
         public string InvoiceNumber { get; set; }
         public string? ExternalUuid { get; set; }
         public string? Cufe { get; set; }
@@ -46,9 +46,8 @@ namespace Poliedro.Eds.Domain.PointOfSale.Entities
 
         public string? WhatsappPhone { get; set; }
 
-        public long? EdsId { get; set; }
-        public long? IsleroId { get; set; }
+        public int? EdsId { get; set; }
+        public int? IsleroId { get; set; }
         public string ProviderTag { get; set; } = "PLEMSI";
     }
-
 }
